perf(calibration): _ece_score y_true skip-cast; 1.27-1.40x int labels
c0006_2f3950b8 @100k (binary cb+xgb + ensembles) profile shows
``_ece_score`` 2002 calls in the honest-diagnostics bootstrap loop
(~10us self / 124us cumtime per call). Same wrapper pattern as
iter595/596/597: an unconditional ``np.asarray(y_true,
dtype=np.float64)`` cast in front of the
``_ece_score_numba_serial`` njit kernel. Bootstrap resamples a
fresh y_true=int slice each iter, so the cast allocates a 100k
float64 buffer per call -- ~800KB x 2002 calls.

The kernel uses ``yi`` only in ``sum_y[b] += yi`` (sum_y is
float64), so mixed-dtype numba dispatch widens at the accumulator,
identical to the explicit upfront cast. ``np.isfinite(yi)`` works
on any numeric dtype. ``pi`` (probabilities, used in ``int(pi *
n_bins)``) stays explicitly float64.

Bench (D:/Temp/bench_iter598_ece_score_cast.py):
  scenario                    OLD (us)   NEW (us)  speedup
  int64 labels @100k            450.51     322.71    1.40x
  int8 labels @100k             438.14     345.16    1.27x
  float64 labels @100k          315.72     317.92    0.99x  (no harm)
  int64 labels @25k             100.28      75.49    1.33x

All bit-equivalent vs the upfront-cast baseline. The float64 case
is essentially flat (no-op cast removed). The 0.99x is within
timing noise.

Tests: +1 ``test_iter598_ece_score_mixed_dtype_bit_equivalent``
pinning bit-equivalence across int64 / int8 / int32 labels.
6 existing tests/calibration/test_pick_best_calibrator.py tests
stay green.

Consistent with the iter595/596/597 pattern: this kernel has 1 op
per element (accumulator add). Per the iter597 rule (skip-cast on
<=3 ops/element, keep cast on 5+) this lands cleanly inside the
"safe" band.

+++ src/mlframe/calibration/policy.py
@@ -118,7 +118,13 @@ def _ece_score(y_true: np.ndarray, p_pred: np.ndarray, n_bins: int = DEFAULT_ECE
     if p.ndim == 2 and p.shape[1] >= 2:
         p = p[:, 1]
     p = np.ascontiguousarray(p.ravel())
-    y = np.ascontiguousarray(np.asarray(y_true, dtype=np.float64).ravel())
+    # iter598: dropped the unconditional ``dtype=np.float64`` cast on
+    # y_true (same pattern as iter595/596/597). The kernel only uses
+    # ``yi`` in ``sum_y[b] += yi`` where sum_y is float64; mixed-dtype
+    # numba dispatch widens at the accumulator, identical to the upfront
+    # cast result. Bench n=100k: int64 1.40x, int8 1.27x, float64 0.99x
+    # (no harm); n=25k int64 (bootstrap typical) 1.33x. Bit-equivalent.
+    y = np.ascontiguousarray(np.asarray(y_true).ravel())
     if p.size == 0 or y.size != p.size:
         return float("nan")
     return _ece_score_numba_serial(y, p, int(n_bins))
